requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
psycopg==3.3.1
psycopg-binary==3.3.1
python-dotenv==1.2.1
sqlparse==0.5.4
djangorestframework-simplejwt
django-filter